+++ packages/iocuak/CHANGELOG.md
@@ -19,6 +19,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 
+## [UNRELEASED]
+
+### Changed
+- `TaskScope` enum is now exposed.
+
+
+
+
 ## v0.0.6 - 2021-10-25
 
 ### Changed